filtering for HR revisions

+++ cms/pages/links.php
@@ -25,8 +25,10 @@
                                 <th>Link Type</th>
                                 <th>Link Tag</th>
                                 <th>Link Description</th>
+                                <?php if($_SESSION['type'] != 3){ ?>
                                 <th>Modify</th>
                                 <th>Delete</th>
+                                <?php } ?>
                             </tr>
                         </thead>
                         <tfoot>
@@ -35,8 +37,10 @@
                                 <th>Link Type</th>
                                 <th>Link Tag</th>
                                 <th>Link Description</th>
+                                <?php if($_SESSION['type'] != 3){ ?>
                                 <th>Modify</th>
                                 <th>Delete</th>
+                                <?php } ?>
                             </tr>
                         </tfoot>
                         <tbody>
@@ -46,12 +50,14 @@
                                     <td><?php echo $link['link_type']; ?></td>
                                     <td><?php echo $link['link_tag']; ?></td>
                                     <td><?php echo $link['link_desc']; ?></td>
-                                    <td><center><button type="button" class="btn bg-green waves-effect" data-toggle="modal" data-target="#edit-link-<?php if($link['link_type'] == 'Link'){?>link<?php } else { ?>file<?php } ?>-modal" data-values="<?php echo htmlspecialchars(json_encode($link)); ?>" onclick="editLink(this);" <?php if($link['status'] == 'Inactive') echo "disabled"; ?>>MODIFY</button></center></td> 
-                                    <?php if($link['status'] == 'Active'){ ?>
-                                        <td><center><button type="button" class="btn bg-red waves-effect" data-type="delete-link" data-id="<?php echo $link['id']; ?>" data-name="<?php echo $link['link_id']; ?>" onclick="alertDesign(this);">DELETE</button></center></td>
-                                    <?php } else { ?>
-                                        <td><center><button type="button" class="btn bg-cyan waves-effect" data-type="reopen-link" data-id="<?php echo $link['id']; ?>" data-name="<?php echo $link['link_id']; ?>" onclick="alertDesign(this);">REACTIVATE</button></center></td>
-                                    <?php } ?>                                    
+                                    <?php if($_SESSION['type'] != 3){ ?>
+                                        <td><center><button type="button" class="btn bg-green waves-effect" data-toggle="modal" data-target="#edit-link-<?php if($link['link_type'] == 'Link'){?>link<?php } else { ?>file<?php } ?>-modal" data-values="<?php echo htmlspecialchars(json_encode($link)); ?>" onclick="editLink(this);" <?php if($link['status'] == 'Inactive') echo "disabled"; ?>>MODIFY</button></center></td> 
+                                        <?php if($link['status'] == 'Active'){ ?>
+                                            <td><center><button type="button" class="btn bg-red waves-effect" data-type="delete-link" data-id="<?php echo $link['id']; ?>" data-name="<?php echo $link['link_id']; ?>" onclick="alertDesign(this);">DELETE</button></center></td>
+                                        <?php } else { ?>
+                                            <td><center><button type="button" class="btn bg-cyan waves-effect" data-type="reopen-link" data-id="<?php echo $link['id']; ?>" data-name="<?php echo $link['link_id']; ?>" onclick="alertDesign(this);">REACTIVATE</button></center></td>
+                                        <?php } ?>
+                                    <?php } ?>
                                 </tr>
                             <?php endforeach; ?>
                         </tbody>
